api: cherry-pick ~fabnev-himnur's invite tweaks
